requirements.txt
langchain-community
langchain
python-dotenv
langchain_core
streamlit